[MOSIP-29726]Update LocationExtnDto.java (#906)

+++ admin/kernel-masterdata-service/src/main/java/io/mosip/kernel/masterdata/dto/getresponse/extn/LocationExtnDto.java
@@ -34,7 +34,7 @@ public class LocationExtnDto extends BaseDto {
 	@JsonDeserialize(using = CustomIntegerDeserializer.class)
 	@Range(min = 0)
 	@ApiModelProperty(value = "hierarchyLevel", required = true, dataType = "java.lang.Integer")
-	private int hierarchyLevel;
+	private short hierarchyLevel;
 
 	@FilterType(types = { FilterTypeEnum.EQUALS, FilterTypeEnum.CONTAINS, FilterTypeEnum.STARTSWITH })
 	@ApiModelProperty(value = "hierarchyName", required = true, dataType = "java.lang.String")
